Made StudyData type definition reusable
